// frontend/src/backend.d.ts
export interface EmailSubscriber {
    email: string;
    timestamp: Time;
}
export type Time = bigint;
export interface FileMetadata {
    path: string;
    size: bigint;
    mimeType: string;
}
// ...
export interface Child {
    name: string;
    birthday?: string;
}
export interface GalleryImage {
    id: string;
    url: string;
    description: string;
    timestamp: bigint;
}
export interface HttpResponse {
    body: Uint8Array;
    headers: Array<HeaderField>;
    streaming_strategy?: StreamingStrategy;
    status_code: number;
}
export interface UnifiedWaiverForm {
    id: string;
    parentEmail: string;
    agreeUnifiedTerms: boolean;
    visitCount: bigint;
    parentPhone: string;
    children: Array<Child>;
    termsAndConditions: string;
    timestamp: bigint;
    parentName: string;
}
// ...
export interface WebsiteContent {
    homepageText: string;
    contactInfo: string;
    aboutText: string;
    operatingHours: string;
}
export interface HttpRequest {
    url: string;
    method: string;
    body: Uint8Array;
    headers: Array<HeaderField>;
}
export interface UserProfile {
    name: string;
}